# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Esport Fantasy League API
  description: |
    Cette documentation décrit les points d'entrée (endpoints) de l'API Gateway du projet **Esport Fantasy League**.

    L'architecture est structurée autour d'un **API Gateway** (port 3000) qui centralise l'accès et route les requêtes vers les services sous-jacents :
    - **Backend Service** (port 3001) : Authentification, gestion des comptes utilisateurs, gestion des ligues fantasy, gestion des rosters et de la draft fantasy.
    - **Esport Data Adapter** (port 3002) : Centralisation et uniformisation des données réelles issues de l'API Pandascore et des API spécifiques de chaque jeu (League of Legends, Counter-Strike, Rocket League, Valorant).

    Toutes les réponses de l'API sont enveloppées dans un objet de réponse standard (`ApiResponse`) contenant :
    - `statusCode` : Le code HTTP de la réponse.
    - `timestamp` : La date et l'heure ISO du traitement de la requête.
    - `data` : La charge utile (payload) retournée.

    ### Sécurité
    La plupart des routes nécessitent une authentification via un jeton **JWT Bearer**. 
    Pour le renouvellement de session, la route `/auth/refresh` attend le **Refresh Token** également sous forme de Bearer token dans les en-têtes Authorization.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/v1
    description: API Gateway (Version 1 - avec préfixe /api/v1)
security:
  - BearerAuth: []
paths:
...
  /leagues:
    get:
      summary: Lister les ligues de l'utilisateur
      description: Récupère la liste de toutes les ligues de fantasy auxquelles l'utilisateur courant participe.
      tags:
        - Ligues Fantasy
      responses:
        "200":
          description: Liste des ligues retournée avec succès.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseLeagueList"
        "401":
          description: Non authentifié.

    post:
      summary: Créer une ligue de fantasy
      description: Permet à l'utilisateur de créer sa propre ligue fantasy avec ses propres règles (jeux ciblés, taille du roster, cooldown de transfert, etc.).
      tags:
        - Ligues Fantasy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - games
                - rosterSize
                - cooldownDays
              properties:
                name:
                  type: string
                  example: League des Pros 2026
                games:
                  type: array
                  items:
                    $ref: "#/components/schemas/Game"
                  example: ["LEAGUE_OF_LEGENDS", "VALORANT"]
                tournaments:
                  type: array
                  items:
                    type: string
                  example: ["LEC 2026 Summer", "VCT EMEA 2026"]
                rosterSize:
                  type: integer
                  minimum: 1
                  maximum: 10
                  example: 5
                cooldownDays:
                  type: integer
                  minimum: 0
                  example: 1
                onlyCreatorInvites:
                  type: boolean
                  default: false
                  example: false
      responses:
        "201":
          description: Ligue créée avec succès.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseLeague"
        "400":
          description: Données de configuration de ligue invalides.
...
  /leagues/{id}:
    get:
      summary: Détails d'une ligue
      description: Récupère les informations complètes d'une ligue spécifique à partir de son ID (y compris la liste de ses membres).
      tags:
        - Ligues Fantasy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: L'identifiant (ID) de la ligue fantasy.
      responses:
        "200":
          description: Détails de la ligue retournés avec succès.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseLeague"
        "404":
          description: Ligue introuvable.

    delete:
      summary: Supprimer une ligue
      description: Supprime définitivement la ligue (uniquement autorisé pour le créateur de la ligue).
      tags:
        - Ligues Fantasy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: L'identifiant de la ligue à supprimer.
      responses:
        "200":
          description: Ligue supprimée avec succès.
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    example: 200
                  timestamp:
                    type: string
                    format: date-time
                  data:
                    type: string
                    example: "League deleted successfully"
        "403":
          description: Non autorisé à supprimer cette ligue (seul le créateur le peut).
        "404":
          description: Ligue introuvable.
  /leagues/{id}/join:
    post:
      summary: Rejoindre une ligue
      description: Permet à l'utilisateur courant de s'enregistrer en tant que membre d'une ligue à l'aide de son code d'invitation unique.
      tags:
        - Ligues Fantasy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: L'identifiant de la ligue à rejoindre.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - inviteCode
              properties:
                inviteCode:
                  type: string
                  example: INV-A1B2-C3D4
      responses:
        "200":
          description: Ligue rejointe avec succès. Retourne les informations de la ligue.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseLeague"
        "400":
          description: Code d'invitation invalide ou ligue pleine (maxMembers atteint).
        "404":
          description: Ligue introuvable.
  /leagues/{leagueId}/leaderboard:
    get:
      summary: Classement de la ligue
      description: Récupère la liste ordonnée des membres de la ligue avec leurs scores cumulés calculés à partir de leurs rosters passés.
      tags:
        - Ligues Fantasy
      parameters:
        - name: leagueId
          in: path
          required: true
          schema:
            type: string
          description: L'identifiant de la ligue.
      responses:
        "200":
          description: Classement retourné avec succès.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponseLeagueMemberList"
        "404":
          description: Ligue introuvable.

  # ────────────────────────────────────────────────────────────────────────
  # ROSTERS (GAMEPLAY FANTASY)
  # ────────────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Injecter l'Access Token JWT (ou Refresh Token pour la route /auth/refresh) obtenu lors de la connexion.
  schemas:
...
    Game:
      type: string
      enum:
        - LEAGUE_OF_LEGENDS
        - COUNTER_STRIKE
        - ROCKET_LEAGUE
        - VALORANT
      example: LEAGUE_OF_LEGENDS
...
    ApiResponseEnvelope:
      type: object
      required:
        - statusCode
        - timestamp
      properties:
        statusCode:
          type: integer
          example: 200
        timestamp:
          type: string
          format: date-time
          example: "2026-07-03T10:50:41.000Z"
...
    LeagueMember:
      type: object
      required:
        - id
        - totalScore
        - user
      properties:
        id:
          type: string
          example: mem_cuid_1
        totalScore:
          type: number
          example: 184.2
        joinedAt:
          type: string
          format: date-time
        user:
          type: object
          required:
            - id
            - username
          properties:
            id:
              type: string
            username:
              type: string
            avatarUrl:
              type: string
              nullable: true
    League:
      type: object
      required:
        - id
        - name
        - games
        - rosterSize
        - cooldownDays
        - inviteCode
        - maxMembers
        - onlyCreatorInvites
        - createdById
        - members
      properties:
        id:
          type: string
          example: lge_cuid_fantasy1
        name:
          type: string
          example: League des Champions
        games:
          type: array
          items:
            $ref: "#/components/schemas/Game"
        tournaments:
          type: array
          items:
            type: string
          example: ["LEC Summer 2026", "LFL Summer 2026"]
        rosterSize:
          type: integer
          example: 5
        cooldownDays:
          type: integer
          example: 1
        inviteCode:
          type: string
          example: INV-K2L9-M10P
        maxMembers:
          type: integer
          example: 12
        onlyCreatorInvites:
          type: boolean
          example: false
        createdById:
          type: string
          example: usr_cljx7d3a00000j
        members:
          type: array
          items:
            $ref: "#/components/schemas/LeagueMember"
        _count:
          type: object
          properties:
            members:
              type: integer
              example: 4
...
    ApiResponseLeague:
      allOf:
        - $ref: "#/components/schemas/ApiResponseEnvelope"
        - type: object
          required:
            - data
          properties:
            data:
              $ref: "#/components/schemas/League"
    ApiResponseLeagueList:
      allOf:
        - $ref: "#/components/schemas/ApiResponseEnvelope"
        - type: object
          required:
            - data
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/League"
    ApiResponseLeagueMemberList:
      allOf:
        - $ref: "#/components/schemas/ApiResponseEnvelope"
        - type: object
          required:
            - data
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/LeagueMember"
...